moved merkle tree to structure module
