+++ cilantro/models/__init__.py
@@ -1,4 +1,4 @@
 from cilantro.models.transaction import StandardTransaction, StandardTransactionBuilder, TransactionBase
 from cilantro.models.base import ModelBase
-from cilantro.models.consensus import MerkleTree, MerkleSignature, BlockContender
+from cilantro.models.consensus import MerkleSignature, BlockContender
 from cilantro.models.message import Message
