+++ cilantro/models/consensus/__init__.py
@@ -1,1 +1,1 @@
-from cilantro.models.consensus.merkle import MerkleTree, MerkleSignature, BlockContender
+from cilantro.models.consensus.merkle import MerkleSignature, BlockContender
