+++ cilantro/protocol/structures/__init__.py
@@ -1,0 +1,1 @@
+from cilantro.protocol.structures.merkle_tree import MerkleTree
